kafka-consumer/requirements.txt
psycopg2
kafka-python